requirements.txt
google-api-python-client==2.116.0
google-auth-oauthlib==1.2.0
PyYAML==6.0.1
